+++ exam/src/main/java/in/testpress/exam/ui/TestActivity.java
@@ -2,13 +2,16 @@ package in.testpress.exam.ui;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.StringRes;
+import androidx.appcompat.app.AlertDialog;
 import androidx.loader.app.LoaderManager;
 import androidx.loader.content.Loader;
 
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.LinearLayout;
@@ -396,7 +399,7 @@ public class TestActivity extends BaseToolBarActivity implements LoaderManager.L
                     new MultiLanguagesUtil.LanguageSelectionListener() {
                         @Override
                         public void onLanguageSelected() {
-                            startExam(false);
+                            showDisableResumeAttemptAlert();
                         }});
             examDuration.setText(exam.getDuration());
         } else {
@@ -434,6 +437,21 @@ public class TestActivity extends BaseToolBarActivity implements LoaderManager.L
             marksPerQuestionLayout.setVisibility(View.GONE);
             negativeMarksLayout.setVisibility(View.GONE);
         }
+    }
+
+    private void showDisableResumeAttemptAlert() {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this, R.style.TestpressAppCompatAlertDialogStyle);
+        builder.setTitle(R.string.exam_resume_disable_warning_title);
+        builder.setMessage(R.string.exam_resume_disable_warning_description);
+        builder.setPositiveButton("OK", new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                startExam(false);
+            }
+        });
+        builder.setNegativeButton("Cancel", null);
+        AlertDialog dialog = builder.create();
+        dialog.show();
     }
 
     private void endExam() {
